pyproject.toml
[tool.poetry]
name = "cpc-webhooks"
version = "1.1.0"
description = ""
authors = [
    "Jose Salcido <jmsalcidoaguilar@gmail.com>",
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
python-telegram-bot = "^21.1.1"
shopifyapi = "^12.4.0"
django = "^5.0.4"
sentry-sdk = {extras = ["django"], version = "^2.0.1"}
djangorestframework = "^3.14.0"
whitenoise = "^6.6.0"
psycopg2-binary = "^2.9.9"
dj-database-url = "^2.1.0"
django-extensions = "^3.2.3"
celery = {extras = ["redis"], version = "^5.3.6"}
django-celery-beat = "^2.5.0"
logtail-python = "^0.2.10"
pytz = "^2024.1"


[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.13"
black = "^24.1.1"
gunicorn = "^21.2.0"
autoflake = "^2.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
